requirements.txt
line-bot-sdk==3.11.0
flask==3.0.0
pygsheets==2.0.6
firebase-admin==6.5.0
azure-cognitiveservices-speech==1.34.0
azure-ai-translation-text==1.0.0